Rename "kalatori_max" to "kalatorimax" and move js files to another dir
