+++ etc/adminhtml/system.xml
@@ -12,7 +12,7 @@
         xsi:noNamespaceSchemaLocation="urn:magento:module:Magento_Config:etc/system_file.xsd">
     <system>
         <section id="payment">
-            <group id="kalatori_max" translate="label" type="text" sortOrder="88" showInDefault="1" showInWebsite="1" showInStore="1">
+            <group id="kalatorimax" translate="label" type="text" sortOrder="100" showInDefault="1" showInWebsite="1" showInStore="1">
                 <label>Dotpayment via Alzymologist</label>
                 <comment>
                     <![CDATA[Not working? <a href="https://developer.alzymologist.com/docs/issues" target="_blank">View common issues</a> or contact <a href="mailto:support@alzymologist.com">support@alzymologist.com</a>]]>
