+++ view/frontend/layout/checkout_index_index.xml
@@ -32,10 +32,10 @@
                                                         <item name="renders" xsi:type="array">
                                                             <!-- merge payment method renders here -->
                                                             <item name="children" xsi:type="array">
-                                                                <item name="kalatori_max" xsi:type="array">
-                                                                    <item name="component" xsi:type="string">Alzymologist_KalatoriMax/js/run</item>
+                                                                <item name="kalatorimax-payments" xsi:type="array">
+                                                                    <item name="component" xsi:type="string">Alzymologist_KalatoriMax/js/view/payment/kalatorimax</item>
                                                                     <item name="methods" xsi:type="array">
-                                                                         <item name="kalatori_max" xsi:type="array">
+                                                                        <item name="kalatorimax" xsi:type="array">
                                                                             <item name="isBillingAddressRequired" xsi:type="boolean">true</item>
                                                                         </item>
                                                                     </item>
