--- view/frontend/web/js/kalatori_max.js
@@ -1,16 +1,0 @@
-alert('Component_00'); 
-
-define(
-    [
-        'Magento_Checkout/js/view/payment/default'
-    ],
-    function (Component) {
-        'use strict';
-alert('Component'); 
-        return Component.extend({
-            defaults: {
-                template: 'Alzymologist_KalatoriMax/payment/kalatori_max'
-            }
-        });
-    }
-);
