+++ view/frontend/web/js/view/payment/method-renderer/kalatorimax.js
@@ -1,0 +1,13 @@
+define(
+    [
+        'Magento_Checkout/js/view/payment/default'
+    ],
+    function (Component) {
+        'use strict';
+        return Component.extend({
+            defaults: {
+                template: 'Alzymologist_KalatoriMax/payment/kalatorimax'
+            }
+        });
+    }
+);
